requirements.txt
Django==2.1.5
wheel==0.32.3
gunicorn==19.9.0
psycopg2==2.7.7
praw==6.1.1
Pillow==5.4.1
google-cloud-storage==1.14.0
djangorestframework==3.8.2
